+++ apps/datatable.py
@@ -5,6 +5,7 @@ import dash_core_components as dcc
 import dash_table_experiments as dt
 import plotly
 import pandas as pd
+import numpy as np
 
 from app import app
 
@@ -12,12 +13,12 @@ WORKING_FOLDER = '/app/'
 df = pd.read_csv(WORKING_FOLDER + 'jme_results.csv',
                  sep=',',
                  dtype={
-                     'severe_wasting': float,
-                     'wasting': float,
-                     'overweight': float,
-                     'stunting': float,
-                     'underweight': float,
-                     'under5': float})
+                     'severe_wasting': np.float64,
+                     'wasting': np.float64,
+                     'overweight': np.float64,
+                     'stunting': np.float64,
+                     'underweight': np.float64,
+                     'under5': np.float64})
 
 layout = html.Div(children=[
     html.H4(children='Malnutrition Data'),
